Refactor monolithic files into focused modules (52 classes → 24 files) (#31)
* Initial plan

* refactor: Split scenario.py into logical modules (21 → 8 files)


* refactor: Split map.py into logical modules (8 → 3 files)


* refactor: Split suspect.py into logical modules (7 → 3 files)


* refactor: Split chat.py into logical modules (7 → 3 files)


* refactor: Split db models scenario.py into logical modules (9 → 4 files)


* fix: Address code review feedback on docstrings and imports


---------
