+++ app/db/models/scenario/__init__.py
@@ -1,0 +1,17 @@
+"""Scenario database models organized by entity."""
+from .main import Scenario
+from .world import Location, VisibilityRule, AccessRule, RequiredEvidence
+from .suspect import Suspect, SuspectTimeline, SuspectSecret
+from .clue import Clue
+
+__all__ = [
+    "Scenario",
+    "Location",
+    "VisibilityRule",
+    "AccessRule",
+    "RequiredEvidence",
+    "Suspect",
+    "SuspectTimeline",
+    "SuspectSecret",
+    "Clue",
+]
