+++ app/models/schemas/map/common.py
@@ -1,0 +1,8 @@
+"""Common map schemas."""
+from pydantic import BaseModel
+
+
+class PositionSchema(BaseModel):
+    """Position coordinates on the map."""
+    x: int
+    y: int
