+++ app/models/schemas/scenario/common.py
@@ -1,0 +1,9 @@
+"""Common schemas used across scenario components."""
+from datetime import time
+from pydantic import BaseModel
+
+
+class TimeRangeSchema(BaseModel):
+    """Time range with start and end times."""
+    start: time
+    end: time
